django_graphql_crud/requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
psycopg2-binary>=2.9.3
graphene-django>=3.0.0
django-graphql-jwt>=0.3.4
django-cors-headers>=3.11.0
PyJWT>=2.3.0
